Moved the PokemonCollectionDialog instanciation out of the init. Moved the 'MainPokemon(...)' function into the pokemon_functions.py file
